Alert for when movies are badly rated or havent been watched for 2 years.
imported java.sql.date and java.time.localdate to make proper timestamps.
Modified lastviewed across the application and database to accept some form of Date datatype.

used this to create a check for when a movie hasnt been watched for 2 years from current date and if a movie is rated 6 or below.